requirements.txt
Flask==3.1.0
flask_sqlalchemy==3.1.1
flask_wtf==1.2.2
injector==0.22.0
openai==1.75.0
pytest==8.3.5
python-dotenv==1.1.0
SQLAlchemy==2.0.40
WTForms==3.2.1
